Add Helpers class, relocate a bunch of functions to there.
